Moved instructions test to base class
